Rename Main form to ControlPanel
