+++ TGControlPanel/ControlPanel.cs
@@ -1,0 +1,60 @@
+﻿using System;
+using System.Drawing;
+using System.Windows.Forms;
+using TGServiceInterface;
+
+namespace TGControlPanel
+{
+	/// <summary>
+	/// The main <see cref="ControlPanel"/> form
+	/// </summary>
+	partial class ControlPanel : Form
+	{
+		/// <summary>
+		/// Create the control panel. Requires the <see cref="Interface"/> has had it's connection info setup
+		/// </summary>
+		public ControlPanel()
+		{
+			InitializeComponent();
+			if (Interface.VersionMismatch(out string error) && MessageBox.Show(error, "Warning", MessageBoxButtons.OKCancel) == DialogResult.Cancel)
+			{
+				Close();
+				return;
+			}
+			Panels.SelectedIndexChanged += Panels_SelectedIndexChanged;
+			Panels.SelectedIndex += Math.Min(Properties.Settings.Default.LastPageIndex, Panels.TabCount - 1);
+			InitRepoPage();
+			InitBYONDPage();
+			InitServerPage();
+			LoadChatPage();
+			InitStaticPage();
+		}
+
+		private void Main_Resize(object sender, EventArgs e)
+		{
+			Panels.Location = new Point(10, 10);
+			Panels.Width = ClientSize.Width - 20;
+			Panels.Height = ClientSize.Height - 20;
+		}
+
+		private void Panels_SelectedIndexChanged(object sender, EventArgs e)
+		{
+			switch (Panels.SelectedIndex)
+			{
+				case 0: //repo
+					PopulateRepoFields();
+					break;
+				case 1: //byond
+					UpdateBYONDButtons();
+					break;
+				case 2: //scp
+					LoadServerPage();
+					break;
+				case 3: //chat
+					LoadChatPage();
+					break;
+			}
+			Properties.Settings.Default.LastPageIndex = Panels.SelectedIndex;
+		}
+	}
+}
