+++ TGControlPanel/ControlPanel.resx
